+++ Changes.md
@@ -5,6 +5,7 @@ Note: development is tracked on the [`develop` branch](https://github.com/chmp/i
 - `develop`:
     - Use markdown for module documentation
     - Updated documentation (updated doc strings for run, ipytest magic)
+    - `ipytest.clean()` now expects a module as its argument
 - `0.12.0`:
     - Re-add the `raise_on_error` config option
     - Return the `exit_code` from `ipyest.run()`